ecommerceBackendAPI/requirements.txt
Django==4.2.4
django_environ==0.11.2
djangorestframework==3.14.0
environ==1.0
PyJWT==2.8.0
pyotp==2.9.0
